# openapi.courses.yaml
openapi: 3.0.3
info:
  title: Courses Service API
  version: 2.0.0
servers:
  - url: http://localhost:8081
    description: Local development
paths:
...
  /api/v1/courses:
    get:
      summary: List courses with filtering, sorting and pagination
      operationId: listCourses
      tags: [courses]
      parameters:
        - name: category
          in: query
          description: Filter by category names (comma-separated)
          schema:
            type: string
            example: Фотография,Дизайн
        - name: level
          in: query
          schema:
            type: string
            enum: [Любой, Новичкам, Продвинутый]
        - name: price_min
          in: query
          schema:
            type: number
            example: 990
        - name: price_max
          in: query
          schema:
            type: number
            example: 3200
        - name: rating_min
          in: query
          schema:
            type: number
            example: 4.0
        - name: sort
          in: query
          schema:
            type: string
            enum: [популярные, новые, по цене, price_desc, по рейтингу]
        - name: search
          in: query
          description: Full-text search by title and author name
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 12
            maximum: 50
      responses:
        "200":
          description: Paginated list of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CourseListItem"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by the Auth Service
...
  schemas:
...
    CourseListItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          example: "Фотография: от новичка до профессионала"
        author:
          type: string
          example: Алексей Петров
        initials:
          type: string
          example: АП
        category:
          type: string
          example: Фотография
        level:
          type: string
          enum: [Любой, Новичкам, Продвинутый]
        price:
          type: number
          example: 2490
        old_price:
          type: number
          nullable: true
          example: 3200
        rating:
          type: number
          format: float
          example: 4.67
        reviews_count:
          type: integer
          example: 3
        students_count:
          type: integer
          example: 42
        hours:
          type: integer
          example: 8
        lessons_count:
          type: integer
          example: 15
        color_1:
          type: string
          example: "#6366f1"
        color_2:
          type: string
          example: "#8b5cf6"
        tag:
          type: string
          nullable: true
          enum: [Бестселлер, Новинка, Популярное]
          example: Бестселлер
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 12
        total:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 4

    # ── Common ───────────────────────────────────────────
...